Centralizes request and response logging
