+++ logging/aws.go
@@ -1,0 +1,29 @@
+package logging
+
+import (
+	"regexp"
+)
+
+// IAM Unique ID prefixes from
+// https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_identifiers.html#identifiers-unique-ids
+var UniqueIDRegex = regexp.MustCompile(`(A3T[A-Z0-9]` +
+	`|ABIA` + // STS service bearer token
+	`|ACCA` + // Context-specific credential
+	`|AGPA` + // User group
+	`|AIDA` + // IAM user
+	`|AIPA` + // EC2 instance profile
+	`|AKIA` + // Access key
+	`|ANPA` + // Managed policy
+	`|ANVA` + // Version in a managed policy
+	`|APKA` + // Public key
+	`|AROA` + // Role
+	`|ASCA` + // Certificate
+	`|ASIA` + // STS temporary access key
+	`)[A-Z0-9]{16,}`)
+
+func MaskAWSAccessKey(field string) string {
+	field = UniqueIDRegex.ReplaceAllStringFunc(field, func(s string) string {
+		return partialMaskString(s, 4, 4) //nolint:gomnd
+	})
+	return field
+}
